Filter PlayerGames by year. Disable ability to change assigned event
* Changing the assigned event would need to move all associated
relations, or duplicate the record and drop all relations. While this
would be nice to have for games, it is complicated and has a lot of
edge cases. Instead we just want to make sure people don’t accidentally
do this.

* Filtering of player games by event was disabled in the past due to in
ability of easily filtering player games by event. I have now added the
eventID to the player game to make this possible. For any current
events, the player game will need to be resaved for this to take effect.